ai-services/duration-predictor/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
